requirements.txt
asgiref==3.5.2
Django==4.0.5
django-countries==7.3.2
django-environ==0.9.0
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
gunicorn==20.1.0
phonenumbers==8.12.54
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
whitenoise==6.2.0
